move onSuccess from usePutTaskCompletion to the mutation hook itself
